Extract bottom app bar items into widgets
