+++ lib/routes/home/widgets/bottom_actions_bar/bottom_actions_bar.dart
@@ -1,26 +1,15 @@
-import 'dart:async';
-
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:c_breez/bloc/account/account_bloc.dart';
 import 'package:c_breez/bloc/account/account_state.dart';
-import 'package:c_breez/bloc/currency/currency_bloc.dart';
-import 'package:c_breez/bloc/currency/currency_state.dart';
-import 'package:c_breez/bloc/input/input_bloc.dart';
 import 'package:c_breez/bloc/lsp/lsp_bloc.dart';
 import 'package:c_breez/bloc/lsp/lsp_state.dart';
 import 'package:c_breez/l10n/build_context_localizations.dart';
-import 'package:c_breez/models/clipboard.dart';
-import 'package:c_breez/routes/spontaneous_payment/spontaneous_payment_page.dart';
-import 'package:c_breez/theme/theme_provider.dart' as theme;
-import 'package:c_breez/widgets/route.dart';
-import 'package:c_breez/widgets/warning_box.dart';
+import 'package:c_breez/routes/home/widgets/bottom_actions_bar/receive_options_bottom_sheet.dart';
+import 'package:c_breez/routes/home/widgets/bottom_actions_bar/send_options_bottom_sheet.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 import 'bottom_action_item.dart';
-import 'bottom_action_item_image.dart';
-import 'enter_payment_info_dialog.dart';
 
 class BottomActionsBar extends StatelessWidget {
   final GlobalKey firstPaymentItemKey;
@@ -32,217 +21,110 @@ class BottomActionsBar extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final texts = context.texts();
     final actionsGroup = AutoSizeGroup();
 
-    return BottomAppBar(
-      child: SizedBox(
-        height: 60,
-        child: Row(
-          mainAxisSize: MainAxisSize.max,
-          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-          children: [
-            BottomActionItem(
-              onPress: () => _showSendOptions(context),
-              group: actionsGroup,
-              text: texts.bottom_action_bar_send,
-              iconAssetPath: "src/icon/send-action.png",
-            ),
-            Container(
-              width: 64,
-            ),
-            BottomActionItem(
-              onPress: () => _showReceiveOptions(context),
-              group: actionsGroup,
-              text: texts.bottom_action_bar_receive,
-              iconAssetPath: "src/icon/receive-action.png",
-            )
-          ],
-        ),
-      ),
-    );
-  }
+    return BlocBuilder<LSPBloc, LSPState>(
+      builder: (context, lsp) {
+        return BlocBuilder<AccountBloc, AccountState>(
+          builder: (context, account) {
+            final connected =
+                lsp.connectionStatus == LSPConnectionStatus.active &&
+                    account.status == AccountStatus.CONNECTED;
 
-  Future _showSendOptions(BuildContext context) async {
-    final texts = AppLocalizations.of(context)!;
-    final inputBloc = context.read<InputBloc>();
-
-    await showModalBottomSheet(
-      context: context,
-      builder: (context) {
-        return BlocBuilder<LSPBloc, LSPState>(
-          builder: (context, lsp) {
-            return BlocBuilder<AccountBloc, AccountState>(
-              builder: (context, account) {
-                return StreamBuilder<DecodedClipboardData>(
-                  stream: inputBloc.decodedClipboardStream,
-                  builder: (context, snapshot) {
-                    final connected =
-                        lsp.connectionStatus == LSPConnectionStatus.active &&
-                            account.status == AccountStatus.CONNECTED;
-                    return Column(
-                      mainAxisSize: MainAxisSize.min,
-                      children: [
-                        const SizedBox(height: 8.0),
-                        ListTile(
-                          enabled: connected,
-                          leading: BottomActionItemImage(
-                            iconAssetPath: "src/icon/paste.png",
-                            enabled: connected,
-                          ),
-                          title: Text(
-                            texts.bottom_action_bar_paste_invoice,
-                            style: theme.bottomSheetTextStyle,
-                          ),
-                          onTap: () => _pasteTapped(
-                            context,
-                            inputBloc,
-                            snapshot.data,
-                          ),
-                        ),
-                        Divider(
-                          height: 0.0,
-                          color: Colors.white.withOpacity(0.2),
-                          indent: 72.0,
-                        ),
-                        ListTile(
-                          enabled: connected,
-                          leading: BottomActionItemImage(
-                            iconAssetPath: "src/icon/connect_to_pay.png",
-                            enabled: connected,
-                          ),
-                          title: Text(
-                            texts.bottom_action_bar_connect_to_pay,
-                            style: theme.bottomSheetTextStyle,
-                          ),
-                          onTap: () => _push(context, "/connect_to_pay"),
-                        ),
-                        Divider(
-                          height: 0.0,
-                          color: Colors.white.withOpacity(0.2),
-                          indent: 72.0,
-                        ),
-                        ListTile(
-                          enabled: connected,
-                          leading: BottomActionItemImage(
-                            iconAssetPath: "src/icon/bitcoin.png",
-                            enabled: connected,
-                          ),
-                          title: Text(
-                            texts.bottom_action_bar_send_btc_address,
-                            style: theme.bottomSheetTextStyle,
-                          ),
-                          onTap: () => _push(context, "/withdraw_funds"),
-                        ),
-                        const SizedBox(height: 8.0)
-                      ],
-                    );
-                  },
-                );
-              },
+            return BottomAppBar(
+              child: SizedBox(
+                height: 60,
+                child: Row(
+                  mainAxisSize: MainAxisSize.max,
+                  mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                  children: [
+                    SendOptions(
+                      connected: connected,
+                      firstPaymentItemKey: firstPaymentItemKey,
+                      actionsGroup: actionsGroup,
+                    ),
+                    Container(width: 64),
+                    ReceiveOptions(
+                      account: account,
+                      connected: connected,
+                      firstPaymentItemKey: firstPaymentItemKey,
+                      actionsGroup: actionsGroup,
+                    )
+                  ],
+                ),
+              ),
             );
           },
         );
       },
     );
   }
+}
 
-  Future _showReceiveOptions(
-    BuildContext context,
-  ) {
+class SendOptions extends StatelessWidget {
+  final bool connected;
+  final GlobalKey<State<StatefulWidget>> firstPaymentItemKey;
+  final AutoSizeGroup actionsGroup;
+
+  const SendOptions({
+    Key? key,
+    required this.connected,
+    required this.firstPaymentItemKey,
+    required this.actionsGroup,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
     final texts = context.texts();
-    final themeData = Theme.of(context);
 
-    return showModalBottomSheet(
-      context: context,
-      builder: (context) {
-        return BlocBuilder<LSPBloc, LSPState>(
-          builder: (context, lsp) {
-            return BlocBuilder<AccountBloc, AccountState>(
-              builder: (context, account) {
-                return BlocBuilder<CurrencyBloc, CurrencyState>(
-                  builder: (context, currencyState) {
-                    final connected =
-                        lsp.connectionStatus == LSPConnectionStatus.active &&
-                            account.status == AccountStatus.CONNECTED;
-                    return Column(
-                      mainAxisSize: MainAxisSize.min,
-                      children: [
-                        const SizedBox(height: 8.0),
-                        ListTile(
-                          enabled: connected,
-                          leading: BottomActionItemImage(
-                            iconAssetPath: "src/icon/paste.png",
-                            enabled: connected,
-                          ),
-                          title: Text(
-                            texts.bottom_action_bar_receive_invoice,
-                            style: theme.bottomSheetTextStyle,
-                          ),
-                          onTap: () => _push(context, "/create_invoice"),
-                        ),
-                        account.maxChanReserve == 0
-                            ? const SizedBox(height: 8.0)
-                            : WarningBox(
-                                boxPadding: const EdgeInsets.all(16),
-                                contentPadding: const EdgeInsets.all(8),
-                                child: AutoSizeText(
-                                  texts.bottom_action_bar_warning_balance_title(
-                                    currencyState.bitcoinCurrency.format(
-                                      account.maxChanReserve,
-                                      removeTrailingZeros: true,
-                                    ),
-                                  ),
-                                  maxFontSize:
-                                      themeData.textTheme.subtitle1!.fontSize!,
-                                  style: themeData.textTheme.headline6,
-                                  textAlign: TextAlign.center,
-                                ),
-                              ),
-                      ],
-                    );
-                  },
-                );
-              },
-            );
-          },
-        );
-      },
+    return BottomActionItem(
+      onPress: () => showModalBottomSheet(
+        context: context,
+        builder: (context) {
+          return SendOptionsBottomSheet(
+            connected: connected,
+            firstPaymentItemKey: firstPaymentItemKey,
+          );
+        },
+      ),
+      group: actionsGroup,
+      text: texts.bottom_action_bar_send,
+      iconAssetPath: "src/icon/send-action.png",
     );
   }
+}
 
-  void _pasteTapped(
-    BuildContext context,
-    InputBloc inputBloc,
-    DecodedClipboardData? clipboardData,
-  ) async {
-    Navigator.of(context).pop();
-    if (clipboardData != null) {
-      final data = clipboardData.data;
-      if (clipboardData.type == ClipboardDataType.paymentRequest) {
-        inputBloc.addIncomingInput(data ?? "");
-      } else if (clipboardData.type == ClipboardDataType.nodeID) {
-        Navigator.of(context).push(FadeInRoute(
-          builder: (_) => SpontaneousPaymentPage(data, firstPaymentItemKey),
-        ));
-      }
-    } else {
-      await showDialog(
-        useRootNavigator: false,
+class ReceiveOptions extends StatelessWidget {
+  final AccountState account;
+  final bool connected;
+  final GlobalKey<State<StatefulWidget>> firstPaymentItemKey;
+  final AutoSizeGroup actionsGroup;
+
+  const ReceiveOptions({
+    Key? key,
+    required this.account,
+    required this.connected,
+    required this.firstPaymentItemKey,
+    required this.actionsGroup,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final texts = context.texts();
+    return BottomActionItem(
+      onPress: () => showModalBottomSheet(
         context: context,
-        barrierDismissible: false,
-        builder: (_) => EnterPaymentInfoDialog(
-          context,
-          inputBloc,
-          firstPaymentItemKey,
-        ),
-      );
-    }
-  }
-
-  void _push(BuildContext context, String route) {
-    final navigatorState = Navigator.of(context);
-    navigatorState.pop();
-    navigatorState.pushNamed(route);
+        builder: (context) {
+          return ReceiveOptionsBottomSheet(
+            account: account,
+            connected: connected,
+            firstPaymentItemKey: firstPaymentItemKey,
+          );
+        },
+      ),
+      group: actionsGroup,
+      text: texts.bottom_action_bar_receive,
+      iconAssetPath: "src/icon/receive-action.png",
+    );
   }
 }
